Move `CountWorkflowExecutions`, `Paginator`, `ServerCapabilities` and `ClientContextTrait` into other namespaces. Deprecate old ones.
