+++ src/Client/ScheduleClientInterface.php
@@ -11,6 +11,7 @@ declare(strict_types=1);
 
 namespace Temporal\Client;
 
+use Temporal\Client\Common\Paginator;
 use Temporal\Client\Schedule\Info\ScheduleListEntry;
 use Temporal\Client\Schedule\Schedule;
 use Temporal\Client\Schedule\ScheduleHandle;
